requirements.txt
pymongo==4.11.3
fastapi==0.115.12
starlette==0.46.1
uvicorn==0.22.0
pyyaml==6.0.2
python-dotenv==1.1.0
python-multipart==0.0.20
httpx==0.28.1
pymupdf==1.25.4
pytest==8.3.5
pytest-asyncio==0.26.0
langchain==0.3.21
langsmith==0.3.19
langchain_community==0.3.20
langchain-openai==0.3.11
openai==1.68.2